+++ projects/ngx-launcher/src/lib/components/activate-booster-createapp-nextstep/activate-booster-createapp-nextstep.component.spec.ts
