+++ projects/ngx-launcher/src/lib/components/dependency-editor-step/dependency-editor-step.component.ts
