+++ projects/ngx-launcher/src/lib/components/project-summary-createapp-step/project-summary-createapp-step.component.spec.ts
