+++ projects/ngx-launcher/src/lib/components/release-strategy-createapp-step/release-strategy-createapp-review.component.html
