+++ projects/ngx-launcher/src/lib/components/release-strategy-createapp-step/release-strategy-createapp-step.component.less
