+++ projects/ngx-launcher/src/lib/components/targetenvironment-createapp-step/target-environment-createapp-step.component.less
